requirements.txt
PySide6>=6.6.0
pyserial>=3.5
requests>=2.31.0
